+++ src/rust/src/conversion/mod.rs
@@ -1,6 +1,7 @@
 use std::num::NonZeroUsize;
 use std::str::FromStr;
 
+use crate::lazyframe::PlROptFlags;
 use crate::prelude::*;
 use crate::{PlRDataFrame, PlRDataType, PlRExpr, PlRLazyFrame, PlRSeries, RPolarsErr};
 pub use categorical::PlRCategories;
@@ -1042,5 +1043,37 @@ impl TryFrom<&str> for Wrap<UnicodeForm> {
             _ => return Err("unreachable".to_string()),
         };
         Ok(Wrap(parsed))
+    }
+}
+
+impl TryFrom<ListSexp> for Wrap<PlROptFlags> {
+    type Error = savvy::Error;
+
+    fn try_from(list: ListSexp) -> Result<Self, savvy::Error> {
+        let opts = PlROptFlags::empty()?;
+        let _ = list.names_iter().map(|x| {
+            let elem_name: &str = x;
+            let elem_value = list.get(x).unwrap().into_typed();
+            let elem_value = match elem_value {
+                TypedSexp::Logical(s) => {
+                    let vec = s.to_vec().to_owned();
+                    *vec.first().unwrap()
+                }
+                _ => unreachable!(),
+            };
+
+            match elem_name {
+                "predicate_pushdown" => opts.set_predicate_pushdown(elem_value),
+                "projection_pushdown" => opts.set_projection_pushdown(elem_value),
+                "simplify_expression" => opts.set_simplify_expression(elem_value),
+                "slice_pushdown" => opts.set_slice_pushdown(elem_value),
+                "comm_subplan_elim" => opts.set_comm_subplan_elim(elem_value),
+                "comm_subexpr_elim" => opts.set_comm_subexpr_elim(elem_value),
+                "cluster_with_columns" => opts.set_cluster_with_columns(elem_value),
+                "check_order_observe" => opts.set_check_order_observe(elem_value),
+                _ => unreachable!(),
+            }
+        });
+        Ok(Wrap(opts))
     }
 }